add: Left/right wave + refactor from Arduino code -- Now one function for left/right starting either up OR down

+++ openwink-mcu-platformio/src/handler/queue.cpp
@@ -1,7 +1,8 @@
 #include "handler/queue.h"
+#include "globals.h"
 
 #include "freertos/FreeRTOS.h"
 
 QueueHandle_t QueueHandler::button_queue = xQueueCreate(10, sizeof(int));
 QueueHandle_t QueueHandler::headlight_input_queue = xQueueCreate(10, sizeof(bool));
-QueueHandle_t QueueHandler::headlight_output_queue = xQueueCreate(10, sizeof(uint8_t));
+QueueHandle_t QueueHandler::headlight_output_queue = xQueueCreate(10, sizeof(int));
